requirements.txt
fastapi[standard]
uvicorn
pandas
python-multipart
psutil
GPUtil
s3fs
soundfile 
librosa
minio
datasets>=2.20.0
datasets[audio]>=2.20.0
accelerate 
evaluate
jiwer 
tensorboard 
gradio
transformers==4.45.2
bitsandbytes
peft
wandb
ctranslate2==4.4.0
faster-whisper
numpy
pandas
pyyaml
jupyterlab